Films page PoC done.
