+++ src/routes/menu.svelte
@@ -85,10 +85,10 @@
 		<a href="/" on:click={handleMenuClose}>
 			<p>home</p>
 		</a>
-		<a href="films" on:click={handleMenuClose}>
+		<a href="/films" on:click={handleMenuClose}>
 			<p>films</p>
 		</a>
-		<a href="about" on:click={handleMenuClose}>
+		<a href="/about" on:click={handleMenuClose}>
 			<p>about</p>
 		</a>
 	</div>
